% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackChanges.R
\name{applyChanges}
\alias{applyChanges}
\title{Replace user's file with diff markup}
\usage{
applyChanges(origPath, newPath)
}
\arguments{
\item{origPath}{The path to the file being edited}

\item{newPath}{The path to the copy of the inital state of the file.}
}
\description{
Replace user's file with diff markup
}
\keyword{internal}
