% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackChanges.R
\name{checkTmpPath}
\alias{checkTmpPath}
\title{Fix problems with temp dir path}
\usage{
checkTmpPath(tmpfilePath)
}
\arguments{
\item{tmpfilePath}{The path to fix}
}
\description{
Fix problems with temp dir path
}
\keyword{internal}
