% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_to_markup.R
\name{diff_to_markup}
\alias{diff_to_markup}
\title{Create track changes diff}
\usage{
diff_to_markup(before, after)
}
\arguments{
\item{before}{(\code{character} of length 1) The text before the changes.}

\item{after}{(\code{character} of length 1) The text after the changes.}
}
\description{
Compares two strings and creates a diff using the markdown track changes syntax.
}
\keyword{internal}
