% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_to_markup.R
\name{getSesValues}
\alias{getSesValues}
\title{Parse SES strings}
\usage{
getSesValues(sesText)
}
\arguments{
\item{sesText}{(\code{character} of length 1) A SES string}
}
\description{
Parse a SES string returned by \code{\link[diffobj]{ses}} into a list of values.
}
\keyword{internal}
