% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackChanges.R
\name{initializeChanges}
\alias{initializeChanges}
\title{Initialize track changes tmp dir}
\usage{
initializeChanges(path)
}
\arguments{
\item{path}{The path to the users original file to track.}
}
\description{
Initialize track changes tmp dir
}
\keyword{internal}
