% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytrack.R
\name{makeBuildScript}
\alias{makeBuildScript}
\title{Make an R script to render changes}
\usage{
makeBuildScript(siteDir)
}
\arguments{
\item{siteDir}{The directory to save the script in.}
}
\description{
Makes an R script to convert markdown with critic markup to an HTML with track changes.
}
\keyword{internal}
