% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackChanges.R
\name{refreshDiff}
\alias{refreshDiff}
\title{Create a HTML of the diff}
\usage{
refreshDiff(origPath, newPath, outDir)
}
\arguments{
\item{origPath}{The path to the file being edited}

\item{newPath}{The path to the copy of the inital state of the file.}

\item{outDir}{Where to save the "index.html" file with the rendered diff.}
}
\description{
Create a HTML rendering of the diff
}
\keyword{internal}
