% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_changes.R
\name{render_changes}
\alias{render_changes}
\title{Render changes as HTML}
\usage{
render_changes(text = NULL, file = NULL, ...)
}
\arguments{
\item{text}{The markdown as a character vector with a single value (i.e. not
one value per line). Either `text` or `file` must be used, but not both.}

\item{file}{The file containing the markdown to be rendered. Either `text` or
`file` must be used, but not both.}

\item{...}{Passed to \code{\link[rmarkdown]{render}}}
}
\description{
Makes and HTML rendering of the track changes markup in markdown.
}
\keyword{internal}
