% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_changes.R
\name{render_html_add}
\alias{render_html_add}
\title{Format text as an addition}
\usage{
render_html_add(text)
}
\arguments{
\item{text}{The text to format}
}
\description{
Format text as an addition using HTML
}
\keyword{internal}
