% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_changes.R
\name{render_html_comment}
\alias{render_html_comment}
\title{Format text as a comment}
\usage{
render_html_comment(comment)
}
\arguments{
\item{from}{What the text was before the change}

\item{to}{What the text is after the change}
}
\description{
Format text as a comment using HTML
}
\keyword{internal}
