% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_changes.R
\name{render_html_delete}
\alias{render_html_delete}
\title{Format text as a deletion}
\usage{
render_html_delete(text)
}
\arguments{
\item{text}{The text to format}
}
\description{
Format text as a deletion using HTML
}
\keyword{internal}
