% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_changes.R
\name{render_html_highlight}
\alias{render_html_highlight}
\title{Format text as a substitution}
\usage{
render_html_highlight(text, comment)
}
\arguments{
\item{from}{What the text was before the change}

\item{to}{What the text is after the change}
}
\description{
Format text as a substitution using HTML
}
\keyword{internal}
