% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{trackSubstitute}
\alias{trackSubstitute}
\title{Tracking a substitution}
\usage{
trackSubstitute()
}
\description{
Call this function as an addin to add the markdown track changes output for substitution.
}
